All buttons working, using ParameterDictionary
This GUI is fully functional for running the model.  It does not yet save a run history.

The items in AxesResizeBug were used for a bug report to MATLAB, and do not need to be kept long-term.  The bottom line is that plots (UIAxes) in the interface can resize incorrectly through no fault of the user, apparently when an app has been moved from one MATLAB version to another.  Workaround: delete the axes and recreate.  Fortunately, user callouts are not deleted during this process.