Delete old app versions and clue up progress plot
The progress plot would get out of sync if the "Go" button got pushed before an existing run was completely done.  Belt-and-suspenders: disable the button until the run returns and plot timers is deleted, AND delete any existing timer before starting the plot each time.  (the plot updates are triggered by this timer)